requirements/local.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.1
django-bootstrap-v5==1.0.11
django-ckeditor==6.4.2
django-js-asset==2.0.0
django-model-utils==4.2.0
idna==3.3
outline-vpn-api==2.0.0
Pillow==9.2.0
psycopg2-binary==2.9.3
pytz==2022.1
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.11

